refactor utils.py to lsi3d/utils/helpers.py
