+++ igibson/utils/data_utils/ext_object/scripts/step_1_visual_mesh.py
@@ -7,7 +7,7 @@ script_dir = os.path.dirname(os.path.abspath(__file__))
 utils_dir = os.path.join(script_dir, "../../blender_utils")
 sys.path.append(utils_dir)
 
-from utils import bake_model, clean_unused, export_ig_object, import_obj_folder
+from lsi_3d.utils.helpers import bake_model, clean_unused, export_ig_object, import_obj_folder
 
 #############################################
 # Parse command line arguments
