+++ igibson/utils/data_utils/ext_object/scripts/step_2_merge.py
@@ -6,7 +6,7 @@ import bpy
 script_dir = os.path.dirname(os.path.abspath(__file__))
 utils_dir = os.path.join(script_dir, "../../blender_utils")
 sys.path.append(utils_dir)
-from utils import clean_unused, import_obj_folder
+from lsi_3d.utils.helpers import clean_unused, import_obj_folder
 
 model_id = sys.argv[-3]
 obj_dir = sys.argv[-2]
