+++ lsi_3d/environment/vision_limit_env.py
@@ -12,7 +12,7 @@ from lsi_3d.environment.tracking_env import TrackingEnv
 from lsi_3d.mdp.lsi_mdp import LsiMdp
 from lsi_3d.mdp.state import AgentState, WorldState
 from lsi_3d.utils.functions import find_nearby_open_spaces
-from utils import grid_to_real_coord, quat2euler, real_to_grid_coord
+from lsi_3d.utils.helpers import grid_to_real_coord, quat2euler, real_to_grid_coord
 
 STICKY_ITEMS = False
